tryon_tray/python/requirements.txt
requests
python-dotenv
tqdm
pyjwt
replicate
# Add other dependencies with versions 